Renamed classes and packages related to event list display for better consistency and clarity. Initialized the list in EventListsState to an empty list instead of null.
