+++ src/main/java/interface_adapter/display_event_lists/DisplayEventListsViewModel.java
@@ -1,0 +1,10 @@
+package interface_adapter.display_event_lists;
+
+import interface_adapter.ViewModel;
+
+public class DisplayEventListsViewModel extends ViewModel<DisplayEventListsState> {
+    public DisplayEventListsViewModel() {
+        super("event lists");
+        setState(new DisplayEventListsState());
+    }
+}
